+++ backend/users/urls.py
@@ -1,7 +1,15 @@
 from django.urls import path
-from .views import login_view, logout_view, guest_login_view, guest_login, set_username
+from .views import (
+    login_view,
+    logout_view,
+    guest_login_view,
+    guest_login,
+    set_username,
+    signup_view,
+)
 
 urlpatterns = [
+    path("signup/", signup_view),
     path("login/", login_view),
     path("logout/", logout_view),
     path("guest/", guest_login_view),